requirements.txt
streamlit>=1.24.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-docx>=0.8.11
matplotlib>=3.7.0
seaborn>=0.12.0
requests>=2.31.0
cdsapi>=0.7.2